% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_label_utils.R
\name{filter_regions_by_size}
\alias{filter_regions_by_size}
\title{Filter regions by pixel count}
\usage{
filter_regions_by_size(labels, min_size = 30L, max_size = Inf)
}
\arguments{
\item{labels}{Matrix of integer labels.}

\item{min_size}{Minimum region size to keep.}

\item{max_size}{Maximum region size to keep.}
}
\value{
Relabeled matrix containing only regions within the requested size range.
}
\description{
Filter regions by pixel count
}
