% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_scale.R
\name{median_scale}
\alias{median_scale}
\title{Center a spectrum by its median}
\usage{
median_scale(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector centered on the median of \code{x}.
}
\description{
Center a spectrum by its median
}
