% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_map.R
\name{plot_region_map}
\alias{plot_region_map}
\title{Plot a segmented region map with dissolved Voronoi-style polygons}
\usage{
plot_region_map(
  cluster_data,
  palette = "magma",
  border_color = "black",
  border_linewidth = 0.8,
  background_color = "black"
)
}
\arguments{
\item{cluster_data}{List containing a \code{cluster_map} matrix.}

\item{palette}{Viridis option name or a vector of colors.}

\item{border_color}{Boundary line color.}

\item{border_linewidth}{Boundary line width.}

\item{background_color}{Background fill color.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a segmented region map with dissolved Voronoi-style polygons
}
