% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_label_utils.R
\name{reindex_regions}
\alias{reindex_regions}
\title{Reindex positive region labels to a compact sequence}
\usage{
reindex_regions(labels)
}
\arguments{
\item{labels}{Matrix of integer labels.}
}
\value{
Matrix with positive labels remapped to \code{1:n}.
}
\description{
Reindex positive region labels to a compact sequence
}
