% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagui-package.R
\docType{package}
\name{sagui-package}
\alias{sagui}
\alias{sagui-package}
\title{sagui}
\description{
Photometric analysis tools for spectral cubes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rafaelsdesouza.com.br/sagui/}
  \item \url{https://github.com/RafaelSdeSouza/sagui}
  \item Report bugs at \url{https://github.com/RafaelSdeSouza/sagui/issues}
}

}
\author{
\strong{Maintainer}: Rafael S. de Souza \email{rd23aag@herts.ac.uk}

}
\keyword{internal}
