% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starlet_mask.R
\name{starlet_mask}
\alias{starlet_mask}
\title{Starlet (à trous) decomposition}
\usage{
starlet_mask(img, J = 5)
}
\arguments{
\item{img}{2D numeric matrix}

\item{J}{number of scales (>=1)}
}
\value{
object of class \code{sagui_starlet}: list(w = list w1..wJ, cJ = coarse)
}
\description{
Starlet (à trous) decomposition
}
