% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_kernel.R
\name{upsample_kernel}
\alias{upsample_kernel}
\title{Insert zeros between kernel taps}
\usage{
upsample_kernel(k, step)
}
\arguments{
\item{k}{base kernel (numeric)}

\item{step}{integer upsampling factor (>=1)}
}
\value{
upsampled kernel
}
\description{
Insert zeros between kernel taps
}
