% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagui-package.R
\docType{package}
\name{sagui-package}
\alias{sagui}
\alias{sagui-package}
\title{sagui}
\description{
Photometric analysis tools for spectral cubes.
}
\author{
\strong{Maintainer}: Rafael S. de Souza \email{rd23aag@herts.ac.uk}

}
\keyword{internal}
